requirements.txt
asgiref==3.8.1
Django==4.0.10
djangorestframework==3.13.1
gunicorn==20.1.0
pytz==2024.1
setuptools==69.5.1
sqlparse==0.5.0
tzdata==2024.1
whitenoise==6.0.0
